Implement network module for the client

+++ chatroom/src/lib.rs
@@ -3,4 +3,4 @@ pub mod protocol;
 pub mod server;
 pub mod client;
 
-pub const DEFAULT_PORT: u32 = 54321;
+pub const DEFAULT_PORT: u16 = 54321;
